Added the interface for filtering. Cosine Similarity filter working now.
