+++ src/main/scala/Factories/AuxiliaryDataFilterFactory.scala
@@ -1,15 +1,16 @@
 package Factories
 
 import Factories.FilterType.FilterType
-import Implementations.AuxiliaryDataRetrievers.FpmAuxiliaryFilter
+import Implementations.AuxiliaryDataRetrievers.{CosineSimAuxiliaryFilter, FpmAuxiliaryFilter}
 import Interfaces.IAuxiliaryDataFilter
 import main.DataTypes.Tweet
+import main.Interfaces.IFeatureGenerator
 import org.apache.spark.rdd.RDD
 
 /**
  * Created by saur6410 on 3/17/17.
  */
-class AuxiliaryDataFilterFactory(trainingTweets: RDD[Tweet]) {
+class AuxiliaryDataFilterFactory(trainingTweets: RDD[Tweet], featureGenerator:IFeatureGenerator) {
 	def getAuxiliaryDataFilter(filterType: FilterType): IAuxiliaryDataFilter = {
 		filterType match {
 			case FilterType.FpmFilter =>
@@ -17,7 +18,10 @@ class AuxiliaryDataFilterFactory(trainingTweets: RDD[Tweet]) {
 					val filter = new FpmAuxiliaryFilter(trainingTweets)
 					filter
 				}
-			case FilterType.CosineSim => null
+			case FilterType.CosineSim => {
+				val positiveTrainingTweets = trainingTweets.filter(trainingTweet => trainingTweet.label == 1.0)
+				new CosineSimAuxiliaryFilter(positiveTrainingTweets,featureGenerator)
+			}
 			case FilterType.Wmd => null
 		}
 	}
